// day_4/utils/types.ts
export interface InputProps extends TextInputProps {
    icon?: React.ReactNode
    containerStyle?: ViewStyle
    inputStyle?: TextStyle
    inputRef?: React.RefObject<TextInput>
    fontSize?: number
    height?: number
    required?: boolean
    validator?: (value: string) => string | null
}